Fixed Lua state creation before debug helper is injected
When attach to x64 UWP process was added in 5078d758568016f5010d8ef2826c27659cbc8a0a, an incorrect test for helper injection was added.
Helper injection is asynchronous and Lua state might be created before it is completed.
Thread suspend and wait was implemented, but when injection to UWP process failed, thread was suspended indefinitely.
Injection was fixed for x64 UWP and to avoid failure in x86 UWP a new check as added.
Now that we have x86 UWP injection implemented, we can remove that check.